Sweet - over that previous painful hump.  :)
Fixup the performance issue from last commit.  Main problem: when drawing a specific part that had steps, was drawing the complete CSI for that part - *from each step*.  This is why all hidden / buffered parts showed up when drawing a submodel.  Also, this meant that a submodel with 10 steps would have the parts in the first step drawn 10 times.  Whoops.

Also, remove the FBO creation that I was naively trying to use on each draw.  It's simply *not* useful.  All drawing was going to the DrawArea's opengl context anyway, because I'm dumb like that.  

Also, change the default page background from white to red, for now, just to ensure we're always able to draw backgrounds...

Whew.  That feels better.