Handle empty steps (step with no parts) and empty page (page with no steps) by warning about them and ignoring.
This rev makes it invalid to try adding a step if there are no pages in the document.  Which means if we're dealing with a brand new non-Lic doc, need to add at least one page, with all steps in it.  Would rather add one page per step, as we're doing now, but that's up to the user (eventually).