Fix step dat generation to that it can handle new steps-in-pages setup.
The whole step rendering in partOGL's renderToPov is a mess, and should really be moved to Step, where it'll be easier to generate the pov with the correct CSI size.  Because generating each step at the full model size is terribly wasteful, and doesn't line up with what openGL rendering expects.