Big changes, a little progress.
Move a bunch of general drawing code out of Draw.py.on_draw into the Page.draw method.  Add a currentPage attribute to Draw, so that it knows what page is currently selected.  Required because we always need to draw a page.

Change the GUI tree to take another column.  Now, column 0 is selection name, 1 is selected node, 2 is selected node's page.  Instead of this approach, I could add a back reference from all Parts/Steps back to Page, so everything knows what page is on... time will tell which approach is best.

Improve the Box class by adding a growBy method, which enlarges the box by a point.  Also define an __add__ method, which allows to add two boxes nicely with +.  Adding two boxes is defined as the union of their rectangles.  Also, Box has a drawAsSelection, which will draw the box as a selection box, ie, a black and white dashed box.  Handy for, well, drawing a selection box around whatever is selected.  We could make this arbitrarily nice and fancy, like maybe a mostly transparent light blue fill over top or something.

Add a boundingBox() method to Page, Step, PartOGL and Part.  This method returns a Box that wholly contains whatever object its called on.  Problems here: Part and PartOGL don't actually have bounding boxes.  What is really meant is specific parts *inside* a PLI or CSI.  Raw PartOGLs should *not* be selectable.  Parts could be selectable, maybe, if clicked on in a CSI.  But that'd be hard for the user to do.  This relates back to the general part selection problem - user needs to select a specific part to be able to displace it, but how to select it?  Given that most selections will be to displace the part *because it's hidden*, they often can't even see it in the CSI.  Maybe a GUI doohiky that expands a part in a PLI to a list, and they can then click on each instance in the list, and the actual part in the CSI will be highlighted?  That'd be neat.  But that's a long way off.

For now, when checking the selected item in Page.draw, check its type, and only select CSI & PLIs.  Of course, to do this CSIs and PLIs actually need to be in the tree to begin with...

Also, remove a bunch more if parenthesis - why was I using those to begin with?  Ugly.  And move some now dead code from Draw to Snippets.

And stop writing such big commit logs.