Add support for rotation steps to opengl output.  That was rather painless - is it possible Lic is slightly better engineered than I keep thinking?  Uh... naaaa.
This adds a new 3D Point class.  Not sure how useful that thing'll be. If it's not used very much when Lic is more mature, ditch it.  If it sticks around, maybe make it a sub class of Point, with one extra parameter?  Inheritance - whodathunk.

Note that only *relative* REL rotation steps (and end steps) are currently supported.  I don't have a handy example of the ADD and ABS types, and don't want to guess at an implementation for them.  Their state is stored, saved & loaded, just ignored.  Implementation should be trivial, but it'll happen when it happens.

One step closer to total Blaster instruction generation.