Convert PLI's layout dictionary to a list, because when sticking these things into a tree, order matters.
Convert the existing Instructions class into a full fledged QAbstractItemModel, which can be thrown into a generic TreeView and displayed.  Works wonderfully... so far.  Had to add a bunch of code to almost every class in here, but it's working quite nicely now, and it's all tied together.

Still need to make the tree do something when clicked on, like change to the selected page, or select the chosen page child, not to mention drag and drop.