Fix tree navigation:
- Whenever something is selected and its tree node is hidden inside its parents, automatically expand that node, to show the selection.

- Fix keyboard tree navigation, so that now the arrow keys work as expected.

TODO: Have a small problem where if you click on a submodel, it's node is expanded automatically and its first page is selected, which is fine.  But if you then use arrow keys to move up or left, selection is automatically forced back to the first page.  Kinda unnatural.  To fix, when a submodel is selected, do not automatically select the first page, select the submodel itself, and just draw its first page.