Simplify page.submodelimage by removing its _row attribute, and allow it to look up its row from the parent page - which is now easily retrieved thanks to the page's children list.
Add a borders list to Page, which keeps track of all the page separators and lines on the page.

Change multi-step page layout so that step 2 is *below* step 1 when there's more than 2 steps.  This gives the nice vertical layout used in regular books.

Add a step separator between each column of steps on a page.  This is not yet saved to the binary blob.  Also not sure if borders will export to the final image, but they should.