Move key handler from the view to the new GraphicsScene subclass, so that Parts can receive key events directly.
Start work on a progress bar for importing models.  This works by making the instructions.loadModel method a generator; it should first calculate a maximum number of increments needed to load the model, based on its size and CSI count, then yield one increment after each step in the init process.  Note that the increment count can come *after* the entire model is loaded, because the time consuming part here is initializing all the sizes and page layouts.