Fix for nasty infinite loops in Callout & Step.getPage().  Need to actually modify the parent being looked up!!
Also, try to get some better AA in the GL output.  At this point, I suspect that's as good as it gets, and problems lie elsewhere, like in polygon / line overlap, or in dumping the GL context to a pixmap, or something.

Also, clean up app settings load / save, use a local .INI file instead of the stupid ugly registry, and save the main splitter position.  Nice.