Add support for changing the default CSI rotation.  This is done in the template CSI's context menu.  Store the default rotation as a class variable (CSI.defaultRotation). Undo / Redo and load / save all work.  Also save default PLI & submodel rotation, for future use.  This nicely cleans up all the rotateToXViews in GLHelpers.
Just need to add a context menu to the submodel & PLI, and hook it up, and all rotation stuff should be good to go.

Also change how the continuous facing layout looks by leaving the template page on its own row.

This commit also includes a much better way to deal with undo / redo buffers in live dialogs... TODO: copy this to the other live dialogs, instead of trying to compose undo macros.