Yet another persistentIndex pile of crap.  Found that if you delete an item that has a persistentIndex, need to remove that index immediately, otherwise that indexes internal pointer goes to hell & back.  Can't wait for LayoutChanged() to come around.
Create a simple way in TreeModel to delete an index given an item, then connect that to Scene's removeItem().

There's gonna be a lot of fallout from this, because now we have to remove an item from the scene *before* killing its parent. But at least the way forward is clear.

Also fixup selection, by giving an easy way to fully select a list of items, in both the scene & the model.  That's more fallout from killing the scene -> model -> scene selection update cycle of slowness.