Push forward on merging Callouts.
Realized that merging two callouts is as easy as removing one and increasing the qty label of the other.  For convenience, store a list of removed callouts in the main callout.  With that 'merged callout' list, can now easily cycle through it, choosing which of the merged callouts to display.

Also, ignore keystrokes in a scene with no page, to avoid crashing.  And clean up tree selection by ignoring stuff that doesn't live in the tree at all, instead of giving those things invalid rows.