First cut at properly merging Page/Step between Submodels.  No undo / redo support, and haven't tested with submodels part lists that don't match exactly yet.
Also fixup mainModel.syncPageNumbers(), so it works properly when called in the middle of editing.  This requires an extra call to synchPageNumbers() after creating the title & part list pages on initial LDrawModel import.

And fix duplicate parts being added to CSIs in submodels.  Don't need to link model parts for all submodels in the submodel dictionary.