Add support for merging all of the selected Steps in one context menu action.
Also clean up Page/Step Submodel cloning: if we don't find a matching part based on filename & color, try ignoring color.  And if we end up with an empty Step, delete it.  With this, cloning two Submodel with different part lists works fine: any part in A not in B is ignored, any part in B not in A will be added to B's last Step.